refactor VM/Opcode module
We've reified the opcodes, each opcode is now an instance of the Opcode
class, with a Proc holding the actual implementation.  VM now just looks
up the appropriate Opcode object from the master
table (Intcode::OPCODES) and then passes itself over to the opcode
implementation to execute.  Note that the Opcode impls are responsible
for incrementing the PC after each instruction.